fixed to/toward issue, troubleshooting asl creation
