+++ renderer/src/EventsContainer/EventsContainer.js
@@ -28,7 +28,7 @@ class EventsContainer extends Component {
 				Events: 
 				<select onChange={(e) => this.setState({ eventToAddFunc: e.target.value })}>
     				<option value="note">note()</option>
-    				<option value="to">to()</option>
+    				<option value="toward">to()</option>
 				</select>
 				<button className="add-event" onClick={this.props.addEvent.bind(this, this.state.eventToAddFunc)}>+</button>
 				{Object.keys(this.props.events).map(
